requirements.txt
aio-pika==8.0.3
aiormq==6.3.4
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.26.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.2.1
#fastapi==0.79.0
fastapi
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
optionaldict==0.1.2
orjson==3.7.8
pamqp==3.2.0
pika==1.3.0
psycopg2==2.9.3
pycryptodome==3.15.0
pydantic==1.9.1
# pydantic
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.10
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
xmltodict==0.13.0
yarl==1.7.2
python-jose
fastapi-users[sqlalchemy]
